Replace using bumpspec by rebase-helper functionality
