+++ packit/utils.py
@@ -289,21 +289,6 @@ def cwd(target):
         os.chdir(curdir)
 
 
-def rpmdev_bumpspec(
-    specfile_path: Path, comment: str, version: str, changelog_file: str = None
-):
-    """
-    wrapper on top of rpmdev-bumpspec utility
-
-    :return:
-    """
-    cmd = ["rpmdev-bumpspec", f"--comment={comment}", f"--new={version}"]
-    if changelog_file:
-        cmd += [f"--file={changelog_file}"]
-    cmd += [str(specfile_path)]
-    run_command(cmd)
-
-
 # TODO: merge this function into parse_git_repo in ogr
 # https://github.com/packit-service/packit/pull/555#discussion_r332871418
 def git_remote_url_to_https_url(inp: str) -> str:
